backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
pydantic==1.8.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==0.19.0
python-multipart
bcrypt==3.2.0